[Setup Sentry] moved script and calling setup from install.sh
Moved the ssentry.py to a file in a subdirectory, so it can be editied
in an editor using syntax highlighting, formatting and autopep
